ci: run tests using the package lifecycle
The jarjar plugin which has been causing issues on CI is attached to the
package phase.
When test is called on CI the package phase is not run, causing the
generated jar to not exist, resulting in test failure.
This moves to using the package phase, which calls tests as a byproduct.